Improve the type of vector in dataframes
In a dataframe, the type of a column is defined by the first cell. If the
cell is empty (equal to NULL), an exception is raised. This patch checks
the vector to determine the type of the vector to allocate. If the vector
is completely empty, a vector of real is allocated and assigned with
NA_REAL constant. (Closes: #2991444).